grpc version update, some Cr feedback around use of BOM, etcd client update for metadata-drivers
