+++ pom.xml
@@ -126,12 +126,11 @@
     <bouncycastle.version>1.60</bouncycastle.version>
     <curator.version>5.1.0</curator.version>
     <dropwizard.version>3.1.0</dropwizard.version>
-    <etcd.version>0.0.2</etcd.version>
+    <etcd.version>0.5.4</etcd.version>
     <freebuilder.version>1.14.9</freebuilder.version>
     <google.code.version>3.0.2</google.code.version>
     <google.errorprone.version>2.1.2</google.errorprone.version>
     <grpc.version>1.35.0</grpc.version>
-    <tomcat-annotations.version>1.35.0</tomcat-annotations.version>
     <guava.version>30.0-jre</guava.version>
     <kerby.version>1.1.1</kerby.version>
     <hadoop.version>2.10.0</hadoop.version>
@@ -473,29 +472,10 @@
       <!-- grpc dependencies -->
       <dependency>
         <groupId>io.grpc</groupId>
-        <artifactId>grpc-all</artifactId>
+        <artifactId>grpc-bom</artifactId>
         <version>${grpc.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>io.grpc</groupId>
-        <artifactId>grpc-netty-shaded</artifactId>
-        <version>${grpc.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>io.grpc</groupId>
-        <artifactId>grpc-protobuf</artifactId>
-        <version>${grpc.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>io.grpc</groupId>
-        <artifactId>grpc-stub</artifactId>
-        <version>${grpc.version}</version>
-      </dependency>
-      <dependency> <!-- necessary for Java 9+ -->
-        <groupId>org.apache.tomcat</groupId>
-        <artifactId>annotations-api</artifactId>
-        <version>${tomcat-annotations.version}</version>
-        <scope>provided</scope>
+        <type>pom</type>
+        <scope>import</scope>
       </dependency>
 
       <!-- rocksdb dependencies -->
